@@ -34,8 +34,7 @@ def analyze_by_access_level(policy_json: dict[str, Any], access_level: str) -> l
     expanded_policy = get_expanded_policy(policy_json)
     requested_actions = get_actions_from_policy(expanded_policy)
     # expanded_actions = determine_actions_to_expand(requested_actions)
-    actions_by_level = remove_actions_not_matching_access_level(requested_actions, access_level)
-    return actions_by_level
+    return remove_actions_not_matching_access_level(requested_actions, access_level)
 
 
 def analyze_statement_by_access_level(statement_json: dict[str, Any], access_level: str) -> list[str]:
@@ -50,5 +49,4 @@ def analyze_statement_by_access_level(statement_json: dict[str, Any], access_lev
     """
     requested_actions = get_actions_from_statement(statement_json)
     expanded_actions = determine_actions_to_expand(requested_actions)
-    actions_by_level = remove_actions_not_matching_access_level(expanded_actions, access_level)
-    return actions_by_level
+    return remove_actions_not_matching_access_level(expanded_actions, access_level)
