@@ -5,13 +5,10 @@ Functions that relate to manipulating files, loading files, and managing filepat
 from __future__ import annotations
 
 import logging
-from typing import TYPE_CHECKING, Any, cast
+from pathlib import Path
+from typing import Any, cast
 
 import yaml
-
-if TYPE_CHECKING:
-    from pathlib import Path
-
 
 logger = logging.getLogger(__name__)
 
@@ -23,7 +20,7 @@ def read_yaml_file(filename: str | Path) -> dict[str, Any]:
     :param filename: name of the yaml file
     :return: dictionary of YAML file contents
     """
-    with open(filename, encoding="utf-8") as yaml_file:
+    with Path(filename).open(encoding="utf-8") as yaml_file:
         try:
             cfg = cast("dict[str, Any]", yaml.safe_load(yaml_file))
         except yaml.YAMLError as exc:
